OpenGL rendering tests
OpenGL could be used to render graphs but THIS IS NOT USED FOR NOW.
Rendering graphs works great but it breaks images and SVG exports.
A Solution could be to use the paintToDevice method to create pixmaps.
SVG might also be generated this way.
At the moment we choose not to focus on this, but this is a very
interesting lead for drawing performance improvements!