Translations added & DarkBlueStyle removed
Translations mechanism almost ready : the menu is not shown at the
moment but could easily be activated in MainWindow.
Translations files (.ts) are prepared in french and english (but not
filled!).
Use build_all_mac.sh to generate the .qm files : they will be embedded
in resources loaded by the application.
TODO : changing language whilst app is running requires to catch a
« LanguageChange » event in all views and to set translated strings
(lot of work to do here…) 